+++ src/main/java/com/example/ame/controller/VeterinarioCanilController.java
@@ -21,7 +21,7 @@ public class VeterinarioCanilController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Tutor> findById(@PathVariable Long id) {
+    public ResponseEntity<Tutor> findById(@PathVariable Integer id) {
         return ResponseEntity.ok(service.findById(id));
     }
