+++ src/main/java/com/example/ame/repository/AtendimentoAmbulatorioRepository.java
@@ -4,5 +4,5 @@ import com.example.ame.model.AtendimentoAmbulatorio;
 import com.example.ame.model.Tutor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AtendimentoAmbulatorioRepository extends JpaRepository<AtendimentoAmbulatorio, Long> {
+public interface AtendimentoAmbulatorioRepository extends JpaRepository<AtendimentoAmbulatorio, Integer> {
 }
