+++ src/main/java/com/example/ame/repository/AtendimentoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AtendimentoRepository extends JpaRepository<Atendimento, Long> {
+public interface AtendimentoRepository extends JpaRepository<Atendimento, Integer> {
 }
