+++ src/main/java/com/example/ame/repository/AtendimentoVeterinarioRepository.java
@@ -4,5 +4,5 @@ import com.example.ame.model.AtendimentoVeterinario;
 import com.example.ame.model.Tutor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AtendimentoVeterinarioRepository extends JpaRepository<AtendimentoVeterinario, Long> {
+public interface AtendimentoVeterinarioRepository extends JpaRepository<AtendimentoVeterinario, Integer> {
 }
