+++ src/main/java/com/example/ame/repository/CastracaoRepository.java
@@ -3,5 +3,5 @@ package com.example.ame.repository;
 import com.example.ame.model.Castracao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CastracaoRepository extends JpaRepository<Castracao, Long> {
+public interface CastracaoRepository extends JpaRepository<Castracao, Integer> {
 }
