+++ src/main/java/com/example/ame/repository/EncaminhamentoRepository.java
@@ -4,5 +4,5 @@ import com.example.ame.model.Castracao;
 import com.example.ame.model.Encaminhamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EncaminhamentoRepository extends JpaRepository<Encaminhamento, Long> {
+public interface EncaminhamentoRepository extends JpaRepository<Encaminhamento, Integer> {
 }
