+++ src/main/java/com/example/ame/repository/ProcedimentosRepository.java
@@ -4,5 +4,5 @@ import com.example.ame.model.Procedimentos;
 import com.example.ame.model.Tutor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProcedimentosRepository extends JpaRepository<Procedimentos, Long> {
+public interface ProcedimentosRepository extends JpaRepository<Procedimentos, Integer> {
 }
