+++ src/main/java/com/example/ame/repository/TutorRepository.java
@@ -3,5 +3,5 @@ package com.example.ame.repository;
 import com.example.ame.model.Tutor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TutorRepository extends JpaRepository<Tutor, Long> {
+public interface TutorRepository extends JpaRepository<Tutor, Integer> {
 }
