+++ src/main/java/com/example/ame/repository/VeterinarioCanilRepository.java
@@ -4,5 +4,5 @@ import com.example.ame.model.Tutor;
 import com.example.ame.model.VeterinarioCanil;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VeterinarioCanilRepository extends JpaRepository<VeterinarioCanil, Long> {
+public interface VeterinarioCanilRepository extends JpaRepository<VeterinarioCanil, Integer> {
 }
